Added path parameter length validation to resource classes.

+++ src/main/java/com/rtds/PacketFilterResource.java
@@ -5,10 +5,12 @@
  */
 package com.rtds;
 
+import com.rtds.event.FilterEvent;
 import com.rtds.jpa.CaptureType;
 import com.rtds.svc.CaptureTypeService;
 import java.util.List;
 import javax.annotation.security.RolesAllowed;
+import javax.enterprise.event.Event;
 import javax.inject.Inject;
 import javax.ws.rs.*;
 
@@ -19,6 +21,8 @@ import javax.ws.rs.*;
 @Path( "/api/filter" )
 public class PacketFilterResource
 {
+    @Inject
+    Event<FilterEvent> filterEvent;
 
     @Inject
     CaptureTypeService captureTypeService;
@@ -27,7 +31,13 @@ public class PacketFilterResource
     @RolesAllowed( "filter_admin" )
     public void addFilter( CaptureType type )
     {
+        if( type.getUrlSuffix().length() > 10 )
+        {
+            throw new IllegalArgumentException( "url_suffix length must be 10 characters or less." );
+        }
+        
         captureTypeService.createOrUpdateCaptureType( type );
+        filterEvent.fire( new FilterEvent( "Filter Added" ) );
     }
 
     @DELETE
@@ -35,11 +45,17 @@ public class PacketFilterResource
     @RolesAllowed( "filter_admin" )
     public void deleteFilter( @PathParam( "url_suffix" ) String url_suffix )
     {
+        if( url_suffix.length() > 10 )
+        {
+            throw new IllegalArgumentException( "url_suffix length must be 10 characters or less." );
+        }
+        
         captureTypeService.deleteCaptureType( url_suffix );
+        filterEvent.fire( new FilterEvent( "Filter Removed" ) );
     }
 
     @GET
-    @RolesAllowed( { "user", "filter_admin" } )
+    @RolesAllowed( { "user", "admin", "filter_admin" } )
     public List<CaptureType> getFilters()
     {
         return captureTypeService.list();
